requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
debugpy==1.3.0
Django==4.2.13
django-debug-toolbar==4.4.2
django-debugger==0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg==3.1.19
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
